+++ CHANGELOG.md
@@ -15,6 +15,7 @@
     * Added the `plusAssign` operator to allow cleaner addition of nodes.
     * Added the `minusAssign` operator to allow cleaner removal of nodes.
     * Added the `putAll` function.
+    * Added the `asSequence` function.
 
 * ### ExperimentalFeature.kt
   * #### File